fix(dbt): cast avg to numeric in sku_days_of_supply for Postgres
Postgres has no round(double precision, integer) — only round(numeric, integer).
round(avg(days_of_supply)::numeric, 1) builds on both Postgres and DuckDB.

+++ dbt/models/marts/sku_days_of_supply.sql
@@ -30,7 +30,7 @@ latest as (
 select
     sku_id,
     count(distinct store_id)                                      as stores_stocked,
-    round(avg(days_of_supply), 1)                                 as avg_days_of_supply,
+    round(avg(days_of_supply)::numeric, 1)                        as avg_days_of_supply,
     sum(inventory)                                                as total_inventory,
     sum(demand)                                                   as total_daily_demand,
     sum(case when risk_status = 'CRITICAL' then 1 else 0 end)     as critical_stores,
